Autonomous-Multi-Agent-Research-Assistant/requirements.txt
streamlit
sentence-transformers
python-dotenv
pydantic
pypdf
reportlab
tavily-python
openai
langchain
pinecone-client